fix(linker): Put the stack in ARM RAM for now.
This is VERY BAD as it grows toward the vector table,
but it frees up space for the heap / static allocation.

Currently we could overwrite the vector table via
sysex anyway if we chose to.  We give the stack as much
space as possible and hope for the best.

Once the app is compiled standalone and loaded into
external DDR, the stack goes back in On-Chip RAM
and we use ARM RAM for interrupt handling.

Signed-off-by: bangcorrupt <bangcorrupt@gmail.com>